Ensures ECS cluster exists before deployment
Modifies the deployment script to check for the existence of the ECS cluster
before proceeding.  This prevents errors if the cluster has not been created
yet.

Also updates the setup documentation to explicitly mention the need to create
the ECS cluster before running the deployment script and updates the
example `run-task` command with example subnet and security group ids.